move pre processing fns to a scaling
