+++ fastats/scaling/__init__.py
@@ -1,0 +1,9 @@
+
+from fastats.scaling.scaling import standard_scale, min_max_scale, rank_data
+
+
+__all__ = [
+    'standard_scale',
+    'min_max_scale',
+    'rank_data',
+]
